# perf(modeling): linear unique points in hullPath2