The function `hullPath2` has a routine to find unique vertices which was implemented with a quadratic array loop. I re-wrote the function to use a map so that it can run in linear time.

With 100 segments this reduces runtime from 2.3ms to 0.9ms (2.5x speedup).
With 1000 segments this reduces runtime from 23ms to 4.5ms (5x speedup).
With 10000 segments this reduces runtime from 750ms to 34ms (22x speedup).
With 50000 segments this reduces runtime from 11500ms to 140ms (80x speedup).
With 100000 segments this reduces runtime from 45000ms to 290ms (150x speedup).

Performance test code:

```js
const jscad = require('@jscad/modeling')
const { geom2, path2 } = jscad.geometries
const { hull } = jscad.hulls
const { circle } = jscad.primitives
const { translate } = jscad.transforms

const main = () => {
  // construct path2
  const circ = circle({ segments: 10000 })
  const outline = geom2.toOutlines(circ)[0]
  const path = path2.fromPoints({closed: true}, outline)

  // construct hull
  const start = performance.now()
  const obj2d = hull(
    translate([17, 21], path),
    translate([7, 0], path),
  )
  console.log(`Hull time ${performance.now() - start} ms`)
  return obj2d
}

module.exports = { main }
```

### All Submissions:

* [x] Have you followed the guidelines in our Contributing document?
* [x] Have you checked to ensure there aren't other open [Pull Requests](../../pulls) for the same update/change?
* [x] Does your submission pass tests?