@platypii 2022-02-23
@z3dev nice point about Map/Set. I changed to Set and the 100k circle example went from 290ms -> 210ms. Awesome! Who needs sleep?